% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmcode.R
\name{lmcode}
\alias{lmcode}
\title{Linear Mixed Codes - an extension of Linear Mixed Models}
\usage{
lmcode(data_list, config = lmcode::extract_config(verbose = F))
}
\arguments{
\item{data_list}{List of data matrices of varying dimensionality. Attempts to find similarities among all datasets with a core structure.}

\item{config}{Configuration parameters (required, default provided)}
}
\value{
Main
}
\description{
Linear Mixed Codes - an extension of Linear Mixed Models
}
