% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmform.R
\name{lmform}
\alias{lmform}
\title{Linear Mixed Forms - an extension of Linear Mixed Models with encoding transformations}
\usage{
lmform(data_list, config = lmform::extract_config(verbose = F))
}
\arguments{
\item{data_list}{List of data matrices of varying dimensionality. Attempts to find similarities among all datasets with a core structure.}

\item{config}{Configuration parameters (required, default provided)}
}
\value{
Main
}
\description{
Linear Mixed Forms - an extension of Linear Mixed Models with encoding transformations
}
